requirements.txt
Django>=2.1.1
Pillow>=5.2.0
argon2-cffi>=18.1.0
django-allauth>=0.35.0
awesome-slugify>=1.6.5
beautifulsoup4==4.6.3
bleach==2.1.4
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
django-appconf==1.0.2
django-crispy-forms==1.7.2
django-widget-tweaks==1.4.3
easy-thumbnails==2.5
html5lib==1.0.1
idna==2.7
olefile==0.46
python-decouple==3.1
pytz==2018.5
requests==2.19.1
six==1.11.0
sorl-thumbnail==12.5.0
urllib3==1.23
webencodings==0.5.1
